+++ CHANGES.rst
@@ -46,7 +46,12 @@ Changelog
    Commit items for NEXT VERSION
    ..............................
 
-   - fix: pyproject.toml section pipenv-unlock (#10)
+   - refactor(cli_dependencies): add logging-strict support (#14)
+   - tests: use pytest fixture logging_strict from package pytest-logging-strict (#14)
+   - refactor: add dependency logging-strict
+   - refactor(dev.in): add dependency pytest-logging-strict
+   - refactor: remove hard coded logging config dict from wreck.constants.LOGGING (#14)
+   - ci: bump action versions
 
 .. scriv-start-here